cafe-management-system1-main/cafemanage/supabase_client.py
from supabase import create_client, Client
from dotenv import load_dotenv
import os

load_dotenv() # Loads SUPABASE_URL and SUPABASE_KEY from .env

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
