cafe-management-system1-main/cafemanage/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
opencv-python>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
bcrypt>=4.0.1   # for password hashing in future
reportlab>=3.6.12
supabase
python-dotenv
